+++ app/Http/Controllers/ParentController.php
@@ -25,7 +25,11 @@ class ParentController extends Controller
      */
     public function create()
     {
-        return view('parents.create');
+        if (auth()->user()) {
+            return view('parents.create');
+        } else {
+            return redirect()->back()->with('error', 'sorry! you are not permitted to perform this action');
+        }
     }
 
     /**
@@ -33,24 +37,28 @@ class ParentController extends Controller
      */
     public function store(StoreparentsRequest $request)
     {
-        $parents = new Parents();
-        $parents->name          =  $request->name;
-        $parents->email         =  $request->email;
-        $parents->password      =  bcrypt($request->password);
-        $parents->telephone     =  $request->telephone;
-        $parents->gender        =  $request->gender;
-        $parents->address       =  $request->address;
-        if ($request->hasFile('image')) {
+        if (auth()->user()) {
+            $parents = new Parents();
+            $parents->name          =  $request->name;
+            $parents->email         =  $request->email;
+            $parents->password      =  bcrypt($request->password);
+            $parents->telephone     =  $request->telephone;
+            $parents->gender        =  $request->gender;
+            $parents->address       =  $request->address;
+            if ($request->hasFile('image')) {
 
-            $image = $request->file('image');
-            $imageName = time().'.'.$image->extension();
-            $image->storeAs('public/images/parents', $imageName);
-            $parents->image = $imageName;
+                $image = $request->file('image');
+                $imageName = time().'.'.$image->extension();
+                $image->storeAs('public/images/parents', $imageName);
+                $parents->image = $imageName;
 
+            }
+            $parents->save();
+
+            return redirect(route('student.login'));
+        } else {
+            return redirect()->back()->with('error', 'sorry! you are not permitted to perform this action');
         }
-        $parents->save();
-
-        return redirect(route('student.login'));
     }
 
     /**
@@ -68,9 +76,13 @@ class ParentController extends Controller
      */
     public function edit(string $id)
     {
-        return view('parents.edit', [
-            'parent' => Parents::findOrFail($id)
-        ]);
+        if (auth()->user()) {
+            return view('parents.edit', [
+                'parent' => Parents::findOrFail($id)
+            ]);
+        } else {
+            return redirect()->back()->with('error', 'sorry! you are not permitted to perform this action');
+        }
     }
 
     /**
@@ -78,26 +90,31 @@ class ParentController extends Controller
      */
     public function update(UpdateparentsRequest $request, string $id)
     {
-        $parents = Parents::findOrFail($id);
-        $parents->name = $request->name;
-        $parents->email = $request->email;
-        $parents->password = bcrypt($request->password);
-        $parents->telephone = $request->telephone;
-        $parents->gender = $request->gender;
-        $parents->address = $request->address;
-         if ($request->hasFile('image')) {
-            // Delete the previous image if it exists
-            if ($parents->image) {
-                Storage::delete('public/images/parents/' . $parents->image);
+        if (auth()->user()) {
+            $parents = Parents::findOrFail($id);
+            $parents->name = $request->name;
+            $parents->email = $request->email;
+            $parents->password = bcrypt($request->password);
+            $parents->telephone = $request->telephone;
+            $parents->gender = $request->gender;
+            $parents->address = $request->address;
+             if ($request->hasFile('image')) {
+                // Delete the previous image if it exists
+                if ($parents->image) {
+                    Storage::delete('public/images/parents/' . $parents->image);
+                }
+                 $image = $request->file('image');
+                $imageName = time().'.'.$image->extension();
+                $image->storeAs('public/images/parents', $imageName);
+                $parents->image = $imageName;
             }
-             $image = $request->file('image');
-            $imageName = time().'.'.$image->extension();
-            $image->storeAs('public/images/parents', $imageName);
-            $parents->image = $imageName;
+             $parents->update();
+             // You can add any additional logic or redirect here if needed
+            return redirect()->back()->with('success', 'Parents updated successfully!');
+
+        } else {
+            return redirect()->back()->with('error', 'sorry! you are not permitted to perform this action');
         }
-         $parents->update();
-         // You can add any additional logic or redirect here if needed
-        return redirect()->back()->with('success', 'Parents updated successfully!');
     }
 
     /**
@@ -105,13 +122,18 @@ class ParentController extends Controller
      */
     public function destroy(string $id)
     {
-        // Delete the associated image if it exists
-        $parents = Parents::findOrFail($id);
-        if ($parents->image) {
-            Storage::delete('public/images/parents/' . $parents->image);
+        if (auth()->user()) {
+            // Delete the associated image if it exists
+            $parents = Parents::findOrFail($id);
+            if ($parents->image) {
+                Storage::delete('public/images/parents/' . $parents->image);
+            }
+            $parents->delete();
+            // You can add any additional logic or redirect here if needed
+            return redirect()->route('parents.index')->with('success', 'Parents deleted successfully!');
+
+        } else {
+            return redirect()->back()->with('error', 'sorry! you are not permitted to perform this action');
         }
-        $parents->delete();
-        // You can add any additional logic or redirect here if needed
-        return redirect()->route('parents.index')->with('success', 'Parents deleted successfully!');
     }
 }
